# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.TrimmingListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.TrimmingRecord'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  handler.createStaffRequest:
    properties:
      email:
        type: string
      job_title_id:
        type: integer
      license_number:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      sort_order:
        type: integer
      staff_role:
        $ref: '#/definitions/model.StaffRole'
    required:
    - email
    - name
    - password
    - staff_role
    type: object
  model.AcquisitionType:
    enum:
    - purchased
    - transferred
    - rescued
    - other
    type: string
    x-enum-varnames:
    - AcquisitionTypePurchase
    - AcquisitionTypeTransfer
    - AcquisitionTypeProtected
    - AcquisitionTypeOther
...
  model.AnimalSpecies:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.BodyWeightUnit:
    enum:
    - Kg
    - g
    type: string
    x-enum-varnames:
    - BodyWeightUnitKg
    - BodyWeightUnitG
...
  model.DangerLevel:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - DangerLevelLow
    - DangerLevelMedium
    - DangerLevelHigh
...
  model.Insurance:
    properties:
      clinic_id:
        type: integer
      contact_phone:
        type: string
      coverage_rate:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.JobTitle:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.MembershipType:
    enum:
    - non_member
    - member
    - deceased
    - transferred
    type: string
    x-enum-varnames:
    - MembershipTypeNonMember
    - MembershipTypeMember
    - MembershipTypeDeceased
    - MembershipTypeOther
...
  model.Owner:
    properties:
      address1:
        type: string
      address2:
        type: string
      birth_date:
        type: string
      clinic_id:
        type: integer
      company:
        type: string
      company_phone:
        type: string
      created_at:
        type: string
      discount_rate:
        type: number
      email:
        type: string
      home_address1:
        type: string
      home_address2:
        type: string
      home_postal_code:
        type: string
      id:
        type: integer
      is_dangerous:
        type: boolean
      membership_type:
        $ref: '#/definitions/model.MembershipType'
      owner_name:
        type: string
      owner_name_kana:
        type: string
      pets:
        description: Relations
        items:
          $ref: '#/definitions/model.Pet'
        type: array
      phone:
        type: string
      postal_code:
        type: string
      remarks:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Pet:
    properties:
      acquisition_type:
        $ref: '#/definitions/model.AcquisitionType'
      animal_species:
        $ref: '#/definitions/model.AnimalSpecies'
      animal_species_id:
        type: integer
      birth_date:
        type: string
      breed:
        type: string
      clinic_id:
        type: integer
      color:
        type: string
      created_at:
        type: string
      danger_level:
        $ref: '#/definitions/model.DangerLevel'
      environment:
        type: string
      food:
        type: string
      gender:
        $ref: '#/definitions/model.PetGender'
      id:
        type: integer
      insurance:
        $ref: '#/definitions/model.Insurance'
      insurance_id:
        type: integer
      last_visit:
        type: string
      name:
        type: string
      neutered_date:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/model.Owner'
        description: Relations
      owner_id:
        type: integer
      pet_name_kana:
        type: string
      pet_number:
        type: string
      phone:
        type: string
      remarks:
        type: string
      status:
        $ref: '#/definitions/model.PetStatus'
      updated_at:
        type: string
      weight:
        type: number
    type: object
  model.PetGender:
    enum:
    - male
    - female
    - unknown
    type: string
    x-enum-varnames:
    - PetGenderMale
    - PetGenderFemale
    - PetGenderUnknown
  model.PetStatus:
    enum:
    - alive
    - deceased
    type: string
    x-enum-varnames:
    - PetStatusAlive
    - PetStatusDeceased
...
  model.ServiceType:
    properties:
      clinic_id:
        type: integer
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  model.Staff:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      job_title:
        allOf:
        - $ref: '#/definitions/model.JobTitle'
        description: Relations
      job_title_id:
        type: integer
      license_number:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      staff_role:
        $ref: '#/definitions/model.StaffRole'
      updated_at:
        type: string
    type: object
...
  model.StaffRole:
    enum:
    - veterinarian
    - nurse
    - trimmer
    - reception
    - manager
    type: string
    x-enum-varnames:
    - StaffRoleVeterinarian
    - StaffRoleNurse
    - StaffRoleTrimmer
    - StaffRoleReception
    - StaffRoleManager
  model.TargetSize:
    enum:
    - small
    - medium
    - large
    - cat
    type: string
    x-enum-varnames:
    - TargetSizeSmall
    - TargetSizeMedium
    - TargetSizeLarge
    - TargetSizeCat
...
  model.TrimmingCourse:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sort_order:
        type: integer
      target_size:
        $ref: '#/definitions/model.TargetSize'
      updated_at:
        type: string
    type: object
  model.TrimmingOption:
    properties:
      clinic_id:
        type: integer
      combinable:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  model.TrimmingRecord:
    properties:
      bt:
        type: string
      bw:
        type: string
      bw_unit:
        $ref: '#/definitions/model.BodyWeightUnit'
      clinic_id:
        type: integer
      completed_image:
        type: string
      course:
        $ref: '#/definitions/model.TrimmingCourse'
      course_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/model.TrimmingOption'
        type: array
      pet:
        allOf:
        - $ref: '#/definitions/model.Pet'
        description: Relations
      pet_id:
        type: integer
      remarks:
        type: string
      staff:
        $ref: '#/definitions/model.Staff'
      staff_id:
        type: integer
      status:
        $ref: '#/definitions/model.TrimmingStatus'
      style_image:
        type: string
      style_request:
        type: string
      updated_at:
        type: string
      used_ribbon:
        type: string
      used_shampoo:
        type: string
      weight:
        type: string
    type: object
  model.TrimmingStatus:
    enum:
    - completed
    - reserved
    - in_progress
    type: string
    x-enum-varnames:
    - TrimmingStatusCompleted
    - TrimmingStatusReserved
    - TrimmingStatusInProgress
...
host: localhost:8080
info:
  contact: {}
  description: 動物病院 電子カルテシステム API（45テーブル・専用マスタテーブル版）
  title: Animal Ekarte API
  version: "2.0"
paths:
...
  /masters/service-types:
    get:
      consumes:
      - application/json
      description: 登録されているサービス種別の一覧を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ServiceType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: サービス種別一覧取得
      tags:
      - Masters
    post:
      consumes:
      - application/json
      description: 新しいサービス種別を作成する
      parameters:
      - description: サービス種別情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ServiceType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ServiceType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: サービス種別作成
      tags:
      - Masters
  /masters/service-types/{id}:
    delete:
      consumes:
      - application/json
      description: 指定IDのサービス種別を削除する
      parameters:
      - description: サービス種別ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: サービス種別削除
      tags:
      - Masters
    put:
      consumes:
      - application/json
      description: 指定IDのサービス種別を更新する
      parameters:
      - description: サービス種別ID
        in: path
        name: id
        required: true
        type: integer
      - description: サービス種別情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ServiceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServiceType'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: サービス種別更新
      tags:
      - Masters
  /masters/staffs:
    get:
      description: クリニックに登録されているスタッフの一覧を返す。roleクエリで職種絞り込み可能。
      parameters:
      - description: 'スタッフ職種でフィルタリング（例: veterinarian, nurse, trimmer, reception, manager）'
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Staff'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: スタッフ一覧取得
      tags:
      - StaffMasters
    post:
      consumes:
      - application/json
      description: 新規スタッフを登録する。スタッフ情報とシステムアカウントをアトミックに作成する。
      parameters:
      - description: スタッフ登録情報
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/handler.createStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Staff'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: スタッフ作成
      tags:
      - StaffMasters
  /masters/staffs/{id}:
    delete:
      description: 指定IDのスタッフを論理削除する。
      parameters:
      - description: スタッフID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: スタッフ削除
      tags:
      - StaffMasters
    put:
      consumes:
      - application/json
      description: 指定IDのスタッフ情報を更新する。
      parameters:
      - description: スタッフID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新するスタッフ情報
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/model.Staff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Staff'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: スタッフ更新
      tags:
      - StaffMasters
...
  /trimmings:
    get:
      consumes:
      - application/json
      description: トリミング記録の一覧をページネーション付きで取得する
      parameters:
      - description: 'ページ番号 (default: 1)'
        in: query
        name: page
        type: integer
      - description: '件数 (1-100, default: 20)'
        in: query
        name: limit
        type: integer
      - description: ペットIDフィルター
        in: query
        name: pet_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TrimmingListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: トリミング一覧取得
      tags:
      - Trimmings
    post:
      consumes:
      - application/json
      description: 新しいトリミング記録を作成する
      parameters:
      - description: トリミング情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TrimmingRecord'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TrimmingRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: トリミング作成
      tags:
      - Trimmings
  /trimmings/{id}:
    get:
      consumes:
      - application/json
      description: 指定IDのトリミング記録を取得する
      parameters:
      - description: トリミングID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrimmingRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: トリミング詳細取得
      tags:
      - Trimmings
    put:
      consumes:
      - application/json
      description: 指定IDのトリミング記録を更新する
      parameters:
      - description: トリミングID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新するトリミング情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TrimmingRecord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrimmingRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: トリミング更新
      tags:
      - Trimmings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"